+++ src/base/resolvers/user-resolver.ts
@@ -2,7 +2,6 @@
 
 import { NDKUser, NDKUserProfile } from '@nostr-dev-kit/ndk';
 import { NostrService } from '../../common/nostr-service';
-import { DEFAULT_PROFILE_IMAGE } from '../../common/constants';
 import { isValidHex, validateNpub, validateNip05 } from '../../common/utils'
 
 export class UserResolver {
